adicionado mensagens flash de acao; correcoes de seguranca.

+++ app/Http/Controllers/ProdutoController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 use App\Models\Produto;
 use Illuminate\Support\Facades\Redirect;
@@ -29,7 +30,7 @@ class ProdutoController extends Controller
         return view('produtos.cadastrar');
     }
     public function store(Request $request){
-
+        if (Auth::user()->nivel == 99) {
         $produto = new Produto;
 
         $produto->NomeProduto = $request->NomeProduto;
@@ -57,9 +58,15 @@ class ProdutoController extends Controller
         $produto->save();
 
         return redirect('cadastrar\produtos')->with('msg', 'Produto cadastrado com SUCESSO');
+    
+    }else{
+        return redirect('/asyudhgashd')->with('msg', 'Você não é autorizado a acessar essapagina');
+    }
+
 
 
  }
+ 
 
 
 
@@ -71,48 +78,81 @@ class ProdutoController extends Controller
     }
 
     public function painel(){
+        if (Auth::user()->nivel == 99) {
+
         $produto = Produto::all();
 
         return view('produtos.painel', ['produto' => $produto]);
     }
+        else {
+            return redirect('/asyudhgashd')->with('msg', 'Você não é autorizado a acessar essapagina');
+        }
+    }
         //deletar
     public function destroy($id){
-        Produto::findOrFail($id)->delete();
-
-
-        return redirect('/p/painel')->with('msg', 'PRODUTO DELETADO COM SUCESSO!');
+        if (Auth::user()->nivel == 99) {
+            Produto::findOrFail($id)->delete();
+            
+            
+            return redirect('/p/painel')->with('msg', 'PRODUTO DELETADO COM SUCESSO!');
+        }
+            else {
+            return redirect('/asyudhgashd')->with('msg', 'Você não é autorizado a acessar essapagina');
+        }
 
     }
 
     //mostrar os deletados
     public function deletados(){
-
+        if (Auth::user()->nivel == 99) {
+         
         $produtos = Produto::onlyTrashed()->get();
 
 
        return view('produtos.deletados', ['produto'=>$produtos]);
 
+    } else {
+            return redirect('/asyudhgashd')->with('msg', 'Você não é autorizado a acessar essapagina');
+        }
+
     }
         //restaurar
     public function restore($deleted_at){
-       $p = Produto::withTrashed()->findOrFail($deleted_at);
-        $p->restore();
-
-        return redirect('/p/painel_deletados')->with('msg', 'Produto restaurado com sucesso!');
+        if (Auth::user()->nivel == 99) {
+            $p = Produto::withTrashed()->findOrFail($deleted_at);
+            $p->restore();
+            
+            return redirect('/p/painel_deletados')->with('msg', 'Produto restaurado com sucesso!');
+            
+        } 
+            else {
+            return redirect('/asyudhgashd')->with('msg', 'Você não é autorizado a acessar essapagina');
+        }
     }
         //mostrar tela de editar
     public function editar($id){
-        $produto = Produto::findOrFail($id);
-
-        return view('produtos.editar', ['produto' => $produto]);
+        if (Auth::user()->nivel == 99) {
+            $produto = Produto::findOrFail($id);
+            
+            return view('produtos.editar', ['produto' => $produto]);
+            
+        } 
+        else {
+            return redirect('/asyudhgashd')->with('msg', 'Você não é autorizado a acessar essapagina');
+        }
     }
         //editar
     public function update(Request $request){
-
-
-        Produto::findOrFail($request->id)->update($request->all());
-
-        return redirect('/p/painel')->with('msg', 'PRODUTO DELETADO COM SUCESSO!');
+        if (Auth::user()->nivel == 99) {
+            
+            
+            Produto::findOrFail($request->id)->update($request->all());
+            
+            return redirect('/p/painel')->with('msg', 'PRODUTO DELETADO COM SUCESSO!');
+        } 
+        else {
+            return redirect('/asyudhgashd')->with('msg', 'Você não é autorizado a acessar essapagina');
+        }
 
     }
 }
